LayoutTests/http/tests/security/xssAuditor/full-block-javascript-link-expected.txt

CONSOLE MESSAGE: line 1: Refused to execute a JavaScript script. Source code of script found within request.

There should be no content in the iframe below:



--------
Frame: 'frame'
--------
